final/commit3

+++ project_movies/src/main/java/com/example/project_movies/dto/CommentDto.java
@@ -13,7 +13,7 @@ public class CommentDto {
 
     private UUID id;
     private String comment;
-    private Integer rating;
+    private double rating;
     private UUID movieId;
